+++ src/containers/FilterContainer.jsx
@@ -1,23 +1,23 @@
 import React, {PropTypes} from 'react'
 
-import WithState from './WithState.jsx'
+import WithState from './../util/withState.js'
 import If from './If.jsx';
 import IfChild from './IfChild.jsx'
 import FilterForm from '../components/FilterForm.jsx';
 import FilterList from '../components/FilterList.jsx'
 
-const FilterContainer = ({state, updateState, filters, addFilterHandler, activateFilterHandler}) => {
+const FilterContainer = ({state, updateState, filters, onAddFilter, onFilterButtonPressed}) => {
     return(
         <div className="filter-container">
           <h3>Filters</h3>
           <If condition={state.formVisibility}>
             <IfChild>
-              <FilterForm onSubmit={filter => {addFilterHandler(filter); updateState({formVisibility: false});}}/>
+              <FilterForm onSubmit={filter => {onAddFilter(filter); updateState({formVisibility: false});}}/>
             </IfChild>
             <IfChild>
               <FilterList
                 filters={ filters }
-                onFilterActiveChange={activateFilterHandler}
+                onFilterActiveChange={onFilterButtonPressed}
               />
               <button onClick={() => updateState({formVisibility: true})}>New filter</button>
             </IfChild>
@@ -32,8 +32,8 @@ FilterContainer.propTypes = {
   }),
   updateState: PropTypes.func.isRequired,
   filters: PropTypes.array.isRequired,
-  addFilterHandler: PropTypes.func,
-  activateFilterHandler: PropTypes.func
+  onAddFilter: PropTypes.func,
+  onFilterButtonPressed: PropTypes.func
 };
 
 export default WithState({formVisibility: false})(FilterContainer);
