+++ src/util/withState.js
@@ -1,16 +1,15 @@
 import React, { Component } from 'react';
 
-import WithState from './WithState.jsx'
-
 const withState = initialState => WrappedComponent => class extends Component {
     constructor(props) {
         super(props);
         this.state = initialState;
+        this.handleStateChange = this.setState.bind(this);
     }
+
+
     render() {
-        return (<WrappedComponent updateState={this.setState.bind(this)}
-                                  state={this.state}
-            {...this.props}/>);
+        return (<WrappedComponent updateState={this.handleStateChange} state={this.state} {...this.props}/>);
     }
 };
